+++ config_baselines.tf
@@ -2,6 +2,7 @@ locals {
   config_topics = [
     module.config_baseline_ap-northeast-1.config_sns_topic,
     module.config_baseline_ap-northeast-2.config_sns_topic,
+    module.config_baseline_ap-northeast-3.config_sns_topic,
     module.config_baseline_ap-south-1.config_sns_topic,
     module.config_baseline_ap-southeast-1.config_sns_topic,
     module.config_baseline_ap-southeast-2.config_sns_topic,
@@ -24,6 +25,7 @@ locals {
 # Reference: https://docs.aws.amazon.com/config/latest/developerguide/gs-cli-prereq.html#gs-cli-create-iamrole
 # --------------------------------------------------------------------------------------------------
 data "aws_iam_policy_document" "recorder_assume_role_policy" {
+  count = var.config_baseline_enabled ? 1 : 0
   statement {
     principals {
       type        = "Service"
@@ -34,14 +36,16 @@ data "aws_iam_policy_document" "recorder_assume_role_policy" {
 }
 
 resource "aws_iam_role" "recorder" {
+  count              = var.config_baseline_enabled ? 1 : 0
   name               = var.config_iam_role_name
-  assume_role_policy = data.aws_iam_policy_document.recorder_assume_role_policy.json
+  assume_role_policy = data.aws_iam_policy_document.recorder_assume_role_policy[0].json
 
   tags = var.tags
 }
 
 # See https://docs.aws.amazon.com/config/latest/developerguide/iamrole-permissions.html
 data "aws_iam_policy_document" "recorder_publish_policy" {
+  count = var.config_baseline_enabled ? 1 : 0
   statement {
     actions   = ["s3:GetBucketAcl", "s3:ListBucket"]
     resources = [local.audit_log_bucket_arn]
@@ -59,20 +63,26 @@ data "aws_iam_policy_document" "recorder_publish_policy" {
   }
 
   statement {
-    actions = ["sns:Publish"]
-
+    actions   = ["sns:Publish"]
     resources = [for topic in local.config_topics : topic.arn if topic != null]
+  }
+
+  statement {
+    actions   = ["kms:Decrypt", "kms:GenerateDataKey"]
+    resources = ["arn:aws:kms:*:${data.aws_caller_identity.current.account_id}:key/${var.config_sns_topic_kms_master_key_id != null ? var.config_sns_topic_kms_master_key_id : ""}"]
   }
 }
 
 resource "aws_iam_role_policy" "recorder_publish_policy" {
+  count  = var.config_baseline_enabled ? 1 : 0
   name   = var.config_iam_role_policy_name
-  role   = aws_iam_role.recorder.id
-  policy = data.aws_iam_policy_document.recorder_publish_policy.json
+  role   = one(aws_iam_role.recorder[*].id)
+  policy = data.aws_iam_policy_document.recorder_publish_policy[0].json
 }
 
 resource "aws_iam_role_policy_attachment" "recorder_read_policy" {
-  role       = aws_iam_role.recorder.id
+  count      = var.config_baseline_enabled ? 1 : 0
+  role       = one(aws_iam_role.recorder[*].id)
   policy_arn = "arn:aws:iam::aws:policy/service-role/AWS_ConfigRole"
 }
 
@@ -88,14 +98,17 @@ module "config_baseline_ap-northeast-1" {
     aws = aws.ap-northeast-1
   }
 
-  enabled                       = contains(var.target_regions, "ap-northeast-1")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "ap-northeast-1")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "ap-northeast-1"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "ap-northeast-1"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_ap-northeast-2" {
@@ -105,14 +118,37 @@ module "config_baseline_ap-northeast-2" {
     aws = aws.ap-northeast-2
   }
 
-  enabled                       = contains(var.target_regions, "ap-northeast-2")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "ap-northeast-2")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "ap-northeast-2"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "ap-northeast-2"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
+}
+
+module "config_baseline_ap-northeast-3" {
+  source = "./modules/config-baseline"
+
+  providers = {
+    aws = aws.ap-northeast-3
+  }
+
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "ap-northeast-3")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
+  s3_bucket_name                = local.audit_log_bucket_id
+  s3_key_prefix                 = var.config_s3_bucket_key_prefix
+  delivery_frequency            = var.config_delivery_frequency
+  sns_topic_name                = var.config_sns_topic_name
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "ap-northeast-3"
+  tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_ap-south-1" {
@@ -122,14 +158,17 @@ module "config_baseline_ap-south-1" {
     aws = aws.ap-south-1
   }
 
-  enabled                       = contains(var.target_regions, "ap-south-1")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "ap-south-1")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "ap-south-1"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "ap-south-1"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_ap-southeast-1" {
@@ -139,14 +178,17 @@ module "config_baseline_ap-southeast-1" {
     aws = aws.ap-southeast-1
   }
 
-  enabled                       = contains(var.target_regions, "ap-southeast-1")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "ap-southeast-1")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "ap-southeast-1"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "ap-southeast-1"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_ap-southeast-2" {
@@ -156,14 +198,17 @@ module "config_baseline_ap-southeast-2" {
     aws = aws.ap-southeast-2
   }
 
-  enabled                       = contains(var.target_regions, "ap-southeast-2")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "ap-southeast-2")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "ap-southeast-2"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "ap-southeast-2"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_ca-central-1" {
@@ -173,14 +218,17 @@ module "config_baseline_ca-central-1" {
     aws = aws.ca-central-1
   }
 
-  enabled                       = contains(var.target_regions, "ca-central-1")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "ca-central-1")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "ca-central-1"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "ca-central-1"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_eu-central-1" {
@@ -190,14 +238,17 @@ module "config_baseline_eu-central-1" {
     aws = aws.eu-central-1
   }
 
-  enabled                       = contains(var.target_regions, "eu-central-1")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "eu-central-1")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "eu-central-1"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "eu-central-1"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_eu-north-1" {
@@ -207,14 +258,17 @@ module "config_baseline_eu-north-1" {
     aws = aws.eu-north-1
   }
 
-  enabled                       = contains(var.target_regions, "eu-north-1")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "eu-north-1")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "eu-north-1"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "eu-north-1"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_eu-west-1" {
@@ -224,14 +278,17 @@ module "config_baseline_eu-west-1" {
     aws = aws.eu-west-1
   }
 
-  enabled                       = contains(var.target_regions, "eu-west-1")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "eu-west-1")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "eu-west-1"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "eu-west-1"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_eu-west-2" {
@@ -241,14 +298,17 @@ module "config_baseline_eu-west-2" {
     aws = aws.eu-west-2
   }
 
-  enabled                       = contains(var.target_regions, "eu-west-2")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "eu-west-2")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "eu-west-2"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "eu-west-2"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_eu-west-3" {
@@ -258,14 +318,17 @@ module "config_baseline_eu-west-3" {
     aws = aws.eu-west-3
   }
 
-  enabled                       = contains(var.target_regions, "eu-west-3")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "eu-west-3")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "eu-west-3"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "eu-west-3"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_sa-east-1" {
@@ -275,14 +338,17 @@ module "config_baseline_sa-east-1" {
     aws = aws.sa-east-1
   }
 
-  enabled                       = contains(var.target_regions, "sa-east-1")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "sa-east-1")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "sa-east-1"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "sa-east-1"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_us-east-1" {
@@ -292,14 +358,17 @@ module "config_baseline_us-east-1" {
     aws = aws.us-east-1
   }
 
-  enabled                       = contains(var.target_regions, "us-east-1")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "us-east-1")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "us-east-1"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "us-east-1"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_us-east-2" {
@@ -309,14 +378,17 @@ module "config_baseline_us-east-2" {
     aws = aws.us-east-2
   }
 
-  enabled                       = contains(var.target_regions, "us-east-2")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "us-east-2")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "us-east-2"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "us-east-2"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_us-west-1" {
@@ -326,14 +398,17 @@ module "config_baseline_us-west-1" {
     aws = aws.us-west-1
   }
 
-  enabled                       = contains(var.target_regions, "us-west-1")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "us-west-1")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "us-west-1"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "us-west-1"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 module "config_baseline_us-west-2" {
@@ -343,14 +418,17 @@ module "config_baseline_us-west-2" {
     aws = aws.us-west-2
   }
 
-  enabled                       = contains(var.target_regions, "us-west-2")
-  iam_role_arn                  = aws_iam_role.recorder.arn
+  enabled                       = var.config_baseline_enabled && contains(var.target_regions, "us-west-2")
+  iam_role_arn                  = one(aws_iam_role.recorder[*].arn)
   s3_bucket_name                = local.audit_log_bucket_id
   s3_key_prefix                 = var.config_s3_bucket_key_prefix
   delivery_frequency            = var.config_delivery_frequency
   sns_topic_name                = var.config_sns_topic_name
-  include_global_resource_types = var.region == "us-west-2"
+  sns_topic_kms_master_key_id   = var.config_sns_topic_kms_master_key_id
+  include_global_resource_types = var.config_global_resources_all_regions ? true : var.region == "us-west-2"
   tags                          = var.tags
+
+  depends_on = [aws_s3_bucket_policy.audit_log]
 }
 
 # --------------------------------------------------------------------------------------------------
@@ -358,7 +436,8 @@ module "config_baseline_us-west-2" {
 # --------------------------------------------------------------------------------------------------
 
 resource "aws_config_config_rule" "iam_mfa" {
-  name = "IAMAccountMFAEnabled"
+  count = var.config_baseline_enabled ? 1 : 0
+  name  = "IAMAccountMFAEnabled"
 
   source {
     owner             = "AWS"
@@ -371,6 +450,7 @@ resource "aws_config_config_rule" "iam_mfa" {
   depends_on = [
     module.config_baseline_ap-northeast-1,
     module.config_baseline_ap-northeast-2,
+    module.config_baseline_ap-northeast-3,
     module.config_baseline_ap-south-1,
     module.config_baseline_ap-southeast-1,
     module.config_baseline_ap-southeast-2,
@@ -389,7 +469,8 @@ resource "aws_config_config_rule" "iam_mfa" {
 }
 
 resource "aws_config_config_rule" "unused_credentials" {
-  name = "UnusedCredentialsNotExist"
+  count = var.config_baseline_enabled ? 1 : 0
+  name  = "UnusedCredentialsNotExist"
 
   source {
     owner             = "AWS"
@@ -404,6 +485,7 @@ resource "aws_config_config_rule" "unused_credentials" {
   depends_on = [
     module.config_baseline_ap-northeast-1,
     module.config_baseline_ap-northeast-2,
+    module.config_baseline_ap-northeast-3,
     module.config_baseline_ap-south-1,
     module.config_baseline_ap-southeast-1,
     module.config_baseline_ap-southeast-2,
@@ -422,7 +504,8 @@ resource "aws_config_config_rule" "unused_credentials" {
 }
 
 resource "aws_config_config_rule" "user_no_policies" {
-  name = "NoPoliciesAttachedToUser"
+  count = var.config_baseline_enabled ? 1 : 0
+  name  = "NoPoliciesAttachedToUser"
 
   source {
     owner             = "AWS"
@@ -441,6 +524,7 @@ resource "aws_config_config_rule" "user_no_policies" {
   depends_on = [
     module.config_baseline_ap-northeast-1,
     module.config_baseline_ap-northeast-2,
+    module.config_baseline_ap-northeast-3,
     module.config_baseline_ap-south-1,
     module.config_baseline_ap-southeast-1,
     module.config_baseline_ap-southeast-2,
@@ -459,7 +543,8 @@ resource "aws_config_config_rule" "user_no_policies" {
 }
 
 resource "aws_config_config_rule" "no_policies_with_full_admin_access" {
-  name = "NoPoliciesWithFullAdminAccess"
+  count = var.config_baseline_enabled ? 1 : 0
+  name  = "NoPoliciesWithFullAdminAccess"
 
   source {
     owner             = "AWS"
@@ -478,6 +563,7 @@ resource "aws_config_config_rule" "no_policies_with_full_admin_access" {
   depends_on = [
     module.config_baseline_ap-northeast-1,
     module.config_baseline_ap-northeast-2,
+    module.config_baseline_ap-northeast-3,
     module.config_baseline_ap-south-1,
     module.config_baseline_ap-southeast-1,
     module.config_baseline_ap-southeast-2,
@@ -500,6 +586,7 @@ resource "aws_config_config_rule" "no_policies_with_full_admin_access" {
 # Only created for the master account.
 # --------------------------------------------------------------------------------------------------
 data "aws_iam_policy_document" "config_organization_assume_role_policy" {
+  count = var.config_baseline_enabled ? 1 : 0
   statement {
     principals {
       type        = "Service"
@@ -510,23 +597,23 @@ data "aws_iam_policy_document" "config_organization_assume_role_policy" {
 }
 
 resource "aws_iam_role" "config_organization" {
-  count = local.is_master_account ? 1 : 0
+  count = var.config_baseline_enabled && local.is_master_account ? 1 : 0
 
   name_prefix        = var.config_aggregator_name_prefix
-  assume_role_policy = data.aws_iam_policy_document.config_organization_assume_role_policy.json
+  assume_role_policy = data.aws_iam_policy_document.config_organization_assume_role_policy[0].json
 
   tags = var.tags
 }
 
 resource "aws_iam_role_policy_attachment" "config_organization" {
-  count = local.is_master_account ? 1 : 0
+  count = var.config_baseline_enabled && local.is_master_account ? 1 : 0
 
   role       = aws_iam_role.config_organization[0].name
   policy_arn = "arn:aws:iam::aws:policy/service-role/AWSConfigRoleForOrganizations"
 }
 
 resource "aws_config_configuration_aggregator" "organization" {
-  count = local.is_master_account ? 1 : 0
+  count = var.config_baseline_enabled && local.is_master_account ? 1 : 0
 
   name = var.config_aggregator_name