+++ modules/alarm-baseline/main.tf
@@ -1,3 +1,6 @@
+data "aws_caller_identity" "current" {}
+data "aws_region" "current" {}
+
 # --------------------------------------------------------------------------------------------------
 # The SNS topic to which CloudWatch alarms send events.
 # --------------------------------------------------------------------------------------------------
@@ -7,7 +10,36 @@ resource "aws_sns_topic" "alarms" {
 
   name = var.sns_topic_name
 
+  kms_master_key_id = var.sns_topic_kms_master_key_id
+
   tags = var.tags
+}
+
+resource "aws_sns_topic_policy" "alarms" {
+  count = var.enabled ? 1 : 0
+  arn   = aws_sns_topic.alarms[0].arn
+
+  policy = data.aws_iam_policy_document.alarms-sns-policy[0].json
+}
+
+data "aws_iam_policy_document" "alarms-sns-policy" {
+  count = var.enabled ? 1 : 0
+
+  statement {
+    actions   = ["sns:Publish"]
+    resources = [aws_sns_topic.alarms[0].arn]
+
+    principals {
+      type        = "Service"
+      identifiers = ["cloudwatch.amazonaws.com"]
+    }
+
+    condition {
+      test     = "ArnLike"
+      variable = "AWS:SourceArn"
+      values   = ["arn:aws:cloudwatch:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:alarm:*"]
+    }
+  }
 }
 
 # --------------------------------------------------------------------------------------------------
@@ -18,7 +50,7 @@ resource "aws_cloudwatch_log_metric_filter" "unauthorized_api_calls" {
   count = var.enabled && var.unauthorized_api_calls_enabled ? 1 : 0
 
   name           = "UnauthorizedAPICalls"
-  pattern        = "{ ($.errorCode = \"*UnauthorizedOperation\") || ($.errorCode = \"AccessDenied*\") }"
+  pattern        = "{(($.errorCode = \"*UnauthorizedOperation\") || ($.errorCode = \"AccessDenied*\")) && (($.sourceIPAddress!=\"delivery.logs.amazonaws.com\") && ($.eventName!=\"HeadBucket\"))}"
   log_group_name = var.cloudtrail_log_group_name
 
   metric_transformation {