+++ securityhub_baselines.tf
@@ -11,7 +11,7 @@ module "securityhub_baseline_ap-northeast-1" {
   providers = {
     aws = aws.ap-northeast-1
   }
-  enabled                          = contains(var.target_regions, "ap-northeast-1")
+  enabled                          = contains(var.target_regions, "ap-northeast-1") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
@@ -25,10 +25,26 @@ module "securityhub_baseline_ap-northeast-2" {
     aws = aws.ap-northeast-2
   }
 
-  enabled                          = contains(var.target_regions, "ap-northeast-2")
+  enabled                          = contains(var.target_regions, "ap-northeast-2") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
+  member_accounts                  = local.securityhub_member_accounts
+}
+
+module "securityhub_baseline_ap-northeast-3" {
+  source = "./modules/securityhub-baseline"
+
+  providers = {
+    aws = aws.ap-northeast-3
+  }
+
+  enabled                          = contains(var.target_regions, "ap-northeast-3") && var.securityhub_enabled
+  enable_cis_standard              = var.securityhub_enable_cis_standard
+  enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
+  enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -39,10 +55,11 @@ module "securityhub_baseline_ap-south-1" {
     aws = aws.ap-south-1
   }
 
-  enabled                          = contains(var.target_regions, "ap-south-1")
+  enabled                          = contains(var.target_regions, "ap-south-1") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -53,10 +70,11 @@ module "securityhub_baseline_ap-southeast-1" {
     aws = aws.ap-southeast-1
   }
 
-  enabled                          = contains(var.target_regions, "ap-southeast-1")
+  enabled                          = contains(var.target_regions, "ap-southeast-1") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -67,10 +85,11 @@ module "securityhub_baseline_ap-southeast-2" {
     aws = aws.ap-southeast-2
   }
 
-  enabled                          = contains(var.target_regions, "ap-southeast-2")
+  enabled                          = contains(var.target_regions, "ap-southeast-2") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -81,10 +100,11 @@ module "securityhub_baseline_ca-central-1" {
     aws = aws.ca-central-1
   }
 
-  enabled                          = contains(var.target_regions, "ca-central-1")
+  enabled                          = contains(var.target_regions, "ca-central-1") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -95,10 +115,11 @@ module "securityhub_baseline_eu-central-1" {
     aws = aws.eu-central-1
   }
 
-  enabled                          = contains(var.target_regions, "eu-central-1")
+  enabled                          = contains(var.target_regions, "eu-central-1") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -109,10 +130,11 @@ module "securityhub_baseline_eu-north-1" {
     aws = aws.eu-north-1
   }
 
-  enabled                          = contains(var.target_regions, "eu-north-1")
+  enabled                          = contains(var.target_regions, "eu-north-1") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -123,10 +145,11 @@ module "securityhub_baseline_eu-west-1" {
     aws = aws.eu-west-1
   }
 
-  enabled                          = contains(var.target_regions, "eu-west-1")
+  enabled                          = contains(var.target_regions, "eu-west-1") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -137,10 +160,11 @@ module "securityhub_baseline_eu-west-2" {
     aws = aws.eu-west-2
   }
 
-  enabled                          = contains(var.target_regions, "eu-west-2")
+  enabled                          = contains(var.target_regions, "eu-west-2") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -151,10 +175,11 @@ module "securityhub_baseline_eu-west-3" {
     aws = aws.eu-west-3
   }
 
-  enabled                          = contains(var.target_regions, "eu-west-3")
+  enabled                          = contains(var.target_regions, "eu-west-3") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -165,10 +190,11 @@ module "securityhub_baseline_sa-east-1" {
     aws = aws.sa-east-1
   }
 
-  enabled                          = contains(var.target_regions, "sa-east-1")
+  enabled                          = contains(var.target_regions, "sa-east-1") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -179,10 +205,11 @@ module "securityhub_baseline_us-east-1" {
     aws = aws.us-east-1
   }
 
-  enabled                          = contains(var.target_regions, "us-east-1")
+  enabled                          = contains(var.target_regions, "us-east-1") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -193,10 +220,11 @@ module "securityhub_baseline_us-east-2" {
     aws = aws.us-east-2
   }
 
-  enabled                          = contains(var.target_regions, "us-east-2")
+  enabled                          = contains(var.target_regions, "us-east-2") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -207,10 +235,11 @@ module "securityhub_baseline_us-west-1" {
     aws = aws.us-west-1
   }
 
-  enabled                          = contains(var.target_regions, "us-west-1")
+  enabled                          = contains(var.target_regions, "us-west-1") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }
 
@@ -221,9 +250,10 @@ module "securityhub_baseline_us-west-2" {
     aws = aws.us-west-2
   }
 
-  enabled                          = contains(var.target_regions, "us-west-2")
+  enabled                          = contains(var.target_regions, "us-west-2") && var.securityhub_enabled
   enable_cis_standard              = var.securityhub_enable_cis_standard
   enable_pci_dss_standard          = var.securityhub_enable_pci_dss_standard
   enable_aws_foundational_standard = var.securityhub_enable_aws_foundational_standard
+  enable_product_arns              = var.securityhub_enable_product_arns
   member_accounts                  = local.securityhub_member_accounts
 }